snapping_info: Fix snap point of L_FLOOR_Q
The value of the snap point location matches the snap point in the game object, but it quits working at large scale. There is another part of that object which is the actual light box and it has a slightly different connection point. That point actually works at larger sacles.