divide: Fix subdividing first segment of divided powerline
Fixes create_point inadvertently renaming an existing object in order to give the caller their desired name.

Code would re-use the _MID name, and the Blender name de-dupe behavior steals that name from an existing object (renaming some other object to have a number suffix). Since we save object names as text, this causes bugs with the existing segments having incorrect references. Most noticeable is the second half of the subdivided segment not getting created (because we look up the saved _MID name by name and get the new midpoint we just created instead of end of this segment which was previously generated with the same name).